+++ specification/transform/delete_transform/DeleteTransformRequest.ts
@@ -32,5 +32,4 @@ export interface Request extends RequestBase {
   query_parameters?: {
     force?: boolean
   }
-  body?: {}
 }
